requirements.txt
anyio==3.7.1
async-timeout==4.0.3
babel==2.17.0
click==8.1.7
courlan==1.3.2
dateparser==1.2.1
Deprecated==1.2.13
fastapi==0.103.2
greenlet==2.0.2
h11==0.14.0
htmldate==1.9.3
idna==3.4
justext==3.0.2
lxml==5.4.0
numpy==1.26.4
packaging==23.2
pandas==1.5.3
pydantic>=2.7.1
pymemcache==3.5.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.2
redis==4.6.0
regex==2024.5.15
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.1
sqlalchemy==1.4.49
starlette==0.27.0
trafilatura==2.0.0
tzdata==2025.2
tzlocal==5.3.1
typing_extensions==4.10.0
uvicorn==0.23.2
wrapt==1.14.1
python-dotenv==1.0.1
tld==0.13.1
lxml_html_clean==0.1.1
beautifulsoup4==4.12.3
langchain>=0.1.16
langchain-aws>=0.0.5
boto3>=1.34.0
botocore>=1.34.0
urllib3<2.0
tqdm
sentence-transformers==2.6.1
scikit-learn==1.4.2

transformers
matplotlib
torch
